Tune the reviews on real projects, add fab output, tolerance analysis and pin automation
Running the reviews over the 18 KiCad demo projects that ship in the image
(tools/review_demos.py, now part of the repo) turned up three problems that the
small test fixture could never show. Across that corpus the output went from
roughly 4600 findings to 467.

  * KiCad's global library tables were only seeded by the container entrypoint,
    so anything that bypassed it - `docker run --entrypoint`, the image used as
    a base - got "the current configuration does not include the library ..."
    once per symbol: 1965 findings of pure noise. Seeding now happens in
    kicad_cli.invoke(), where it cannot be skipped.
  * Symbol.is_power only recognised the `power:` library. KiCad's real invariant
    is the `#` reference prefix, so every GND symbol in the older demos was
    counted as a component with no footprint (784 false findings).
  * net.single_pin fired 656 times. An auto-named net (unconnected-(U1-Pad3)) is
    a dangling wire and still warns; a net the designer named is a deliberate
    spare and is now info. On interf_u the review went from 113 warnings to 2.

Any rule that fires more than COLLAPSE_LIMIT times is folded into one finding
carrying the count and the first examples (util.collapse_findings), so one noisy
rule cannot bury the rest of a report. `--collapse N` tunes it, 0 disables it.

New capabilities:

  * `eda pcb fab` writes the manufacturing package - Gerbers (X2), Excellon
    drill with map and report, pick-and-place, BOM, optional STEP/IPC-2581 -
    zipped with a manifest. Each step fails independently. New skill:
    kicad-fabrication-output.
  * `eda sch bom` exports a grouped bill of materials with a parsed summary.
  * `eda sim montecarlo` resamples component values inside their tolerance
    (normal / uniform / worst case, seeded) and reports mean, stdev, percentiles
    and spread of any measurement, with a histogram and per-trial CSV.
  * `eda sim temperature` runs the deck across a temperature range and reports
    the drift per degree.

Supply chain and hygiene:

  * .github/dependabot.yml for uv.lock and the actions; tools/refresh_pins.py
    plus a weekly workflow for what Dependabot cannot parse (the KiCad image
    digest, the pip and uv bootstrap wheels). The default KiCad release is never
    bumped automatically. `make check-pins` reports drift; KiCad 10.0.5 is now
    pinned as an opt-in.
  * ruff (lint in CI, config in pyproject.toml) and coverage on the container
    job, uploaded as an artifact.

180 tests pass in the container and the smoke script covers every new command.

+++ src/eda_toolkit/spice/measure.py
@@ -179,7 +179,7 @@ def thd(plot: Plot, signal: str, fundamental_hz: float, *, harmonics: int = 7,
     uniform_v = np.interp(uniform_t, time, values)
     spectrum = np.abs(np.fft.rfft(uniform_v)) * 2.0 / n
     bin_width = 1.0 / window
-    fundamental_bin = int(round(fundamental_hz / bin_width))
+    fundamental_bin = round(fundamental_hz / bin_width)
     if fundamental_bin >= len(spectrum):
         raise ValueError("fundamental is above the sampled bandwidth")
     fundamental_amp = float(spectrum[fundamental_bin])
